+++ cmd/auth-login.go
@@ -61,5 +61,13 @@ func init() {
 		codeLoginCmd.GetDocFlag("client-id").Default,
 		codeLoginCmd.GetDocFlag("client-id").Description,
 	)
+
+	// intentionally a string flag to support an empty port which represents a dynamic port
+	codeLoginCmd.Flags().StringP(
+		codeLoginCmd.GetDocFlag("port").Name,
+		codeLoginCmd.GetDocFlag("port").Shorthand,
+		codeLoginCmd.GetDocFlag("port").Default,
+		codeLoginCmd.GetDocFlag("port").Description,
+	)
 	authCmd.AddCommand(&codeLoginCmd.Command)
 }