+++ cmd/tdf-decrypt.go
@@ -63,12 +63,14 @@ func dev_tdfDecryptCmd(cmd *cobra.Command, args []string) {
 	ignoreAllowlist := len(kasAllowList) == 1 && kasAllowList[0] == "*"
 
 	decrypted, err := h.DecryptBytes(
+		c.Context(),
 		bytesToDecrypt,
 		assertionVerification,
 		disableAssertionVerification,
 		sessionKeyAlgorithm,
 		kasAllowList,
 		ignoreAllowlist,
+		nil,
 	)
 	if err != nil {
 		cli.ExitWithError("Failed to decrypt file", err)